adicionando carga horaria

+++ Dtos/EventoDto.cs
@@ -5,7 +5,7 @@ public class EventoDto
     public string Titulo { get; set; }
     public DateTime Data { get; set; }
     public String Local { get; set; }
-    public long HorasComplementares { get; set; }
+    public int HorasComplementares { get; set; }
     public string MatriculaPalestrante { get; set; }
     public string MatriculaOrganizador { get; set; }
 }
